backend/main-api/requirements.txt
flask
pymongo
requests
validators
passlib
Flask-Authlib-Client